AttachDownloader/requirements.txt
# Dependencias para Gmail Downloader
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
